Fix ecto type breaking when no ecto
